backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
chromadb
sentence-transformers
ollama
